+++ tests/integration/conftest.py
@@ -164,3 +164,13 @@ def workspace_with_adaptive_hooks(setup: Setup, tmp_path: Path) -> Workspace:
         url=str(tmp_path / "repo"),
         config_url=str(tmp_path / "configs" / "with-adaptive-hooks"),
     )
+
+
+@pytest.fixture
+def workspace_with_cache_hooks(setup: Setup, tmp_path: Path) -> Workspace:
+    setup(config="with-cache-hooks")
+    return Workspace.clone(
+        workspace_dir=str(tmp_path / "workspace"),
+        url=str(tmp_path / "repo"),
+        config_url=str(tmp_path / "configs" / "with-cache-hooks"),
+    )